backend work for collections near final

+++ packages/server/src/app/api/content/collections/collections.controller.ts
@@ -1,18 +1,40 @@
-import { Controller, UseGuards, Request, Param, Body, Get, Put, Patch } from '@nestjs/common';
+import { Controller, UseGuards, Request, Param, Body, Get, Put, Patch, Query, BadRequestException } from '@nestjs/common';
 
 import { RolesGuard } from '../../../guards';
 import { Roles } from '@pulp-fiction/models/users';
 import { CreateCollection, EditCollection } from '@pulp-fiction/models/collections';
 import { CollectionsService } from '../../../db/collections/collections.service';
+import { isNullOrUndefined } from '../../../util';
 
 @Controller('collections')
 export class CollectionsController {
     constructor(private readonly collsService: CollectionsService) {}
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Get('fetch-user-collections/:pageNum')
-    async fetchUserCollections(@Request() req: any, @Param('pageNum') pageNum: number) {
-        return await this.collsService.getUserCollections(req.user, pageNum);
+    @Get('get-all-collections')
+    async getAllCollections(@Request() req: any, @Query('pageNum') pageNum: number) {
+        if (isNullOrUndefined(pageNum)) {
+            throw new BadRequestException(`This request must include the page number.`);
+        }
+        return await this.collsService.getAllCollections(req.user, pageNum);
+    }
+
+    @UseGuards(RolesGuard([Roles.User]))
+    @Get('get-one-collection')
+    async getOneCollection(@Request() req: any, @Query('collId') collId: string, @Query('getPublic') getPublic: boolean) {
+        if (isNullOrUndefined(collId)) {
+            throw new BadRequestException(`This request requires the collection ID.`);
+        }
+        return await this.collsService.getOneCollection(req.user, collId, getPublic);
+    }
+
+    @UseGuards(RolesGuard([Roles.User]))
+    @Get('get-public-collections')
+    async getPublicCollections(@Request() req: any, @Query('pageNum') pageNum: number) {
+        if (isNullOrUndefined(pageNum)) {
+            throw new BadRequestException(`This request must include the page number.`);
+        }
+        return await this.collsService.getPublicCollections(req.user, pageNum);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
@@ -28,38 +50,56 @@ export class CollectionsController {
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('edit-collection/:collId')
-    async editCollection(@Request() req: any, @Param('collId') collId: string, @Body() collInfo: EditCollection) {
+    @Patch('edit-collection')
+    async editCollection(@Request() req: any, @Query('collId') collId: string, @Body() collInfo: EditCollection) {
+        if (isNullOrUndefined(collId)) {
+            throw new BadRequestException(`This request requires the collection ID.`);
+        }
         return await this.collsService.editCollection(req.user, collId, collInfo);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('delete-collection/:collId')
-    async deleteCollection(@Request() req: any, @Param('collId') collId: string) {
+    @Patch('delete-collection')
+    async deleteCollection(@Request() req: any, @Query('collId') collId: string) {
+        if (isNullOrUndefined(collId)) {
+            throw new BadRequestException(`This request requires the collection ID.`);
+        }
         return await this.collsService.deleteCollection(req.user, collId);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('add-work/:collId/:workId')
-    async addWork(@Request() req: any, @Param('collId') collId: string, @Param('workId') workId: string) {
-        return await this.collsService.addWorkToCollection(req.user, collId, workId);
+    @Patch('add-content')
+    async addContent(@Request() req: any, @Query('collId') collId: string, @Query('contentId') contentId: string) {
+        if (isNullOrUndefined(collId) || isNullOrUndefined(contentId)) {
+            throw new BadRequestException(`This request requires the collection ID and the content ID.`);
+        }
+        return await this.collsService.addToCollection(req.user, collId, contentId);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('remove-work/:collId/:workId')
-    async removeWork(@Request() req: any, @Param('collId') collId: string, @Param('workId') workId: string) {
-        return await this.collsService.removeWorkFromCollection(req.user, collId, workId);
+    @Patch('remove-content')
+    async removeContent(@Request() req: any, @Query('collId') collId: string, @Query('contentId') contentId: string) {
+        if (isNullOrUndefined(collId) || isNullOrUndefined(contentId)) {
+            throw new BadRequestException(`This request requires the collection ID, the collection item ID, and the content ID.`);
+        }
+        return await this.collsService.removeFromCollection(req.user, collId, contentId);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('set-public/:collId')
-    async setPublic(@Request() req: any, @Param('collId') collId: string) {
+    @Patch('set-public')
+    async setPublic(@Request() req: any, @Query('collId') collId: string) {
+        if (isNullOrUndefined(collId)) {
+            throw new BadRequestException(`This request requires the collection ID.`);
+        }
         return await this.collsService.setPublic(req.user, collId);
     }
 
     @UseGuards(RolesGuard([Roles.User]))
-    @Patch('set-private/:collId')
-    async setPrivate(@Request() req: any, @Param('collId') collId: string) {
+    @Patch('set-private')
+    async setPrivate(@Request() req: any, @Query('collId') collId: string) {
+        if (isNullOrUndefined(collId)) {
+            throw new BadRequestException(`This request requires the collection ID.`);
+        }
         return await this.collsService.setPrivate(req.user, collId);
     }
 }
